app/python/requirements.txt
   mysql-connector-python
   numpy
   scikit-learn
